requirements.txt
google-cloud-texttospeech
PyMuPDF
PyQt6